User Class Refactor
Rewrote the user class by splitting it into
three separate classes:
User:- Holds the users data and allows access to it.
Login:- Implements the Login Function
Register:- Registers logged in users.
Data is passed between the three functions by way of an array.
