+++ includes/users/User.php
@@ -71,14 +71,6 @@ class User
     protected $encryptedPasswd = '';
 
     /**
-     * Database connection object
-     *
-     * @var    \g7mzr\db\interfaces\InterfaceDatabaseDriver
-     * @access protected
-     */
-    protected $db  = null;
-
-    /**
      * An array holding the current user's permissions
      *
      * @var    array
@@ -116,7 +108,7 @@ class User
      * @var    string
      * @access protected
      */
-    protected $passwdagemsg = '';
+    protected $passwdAgeMsg = '';
 
     /**
      * The Default Style Directory
@@ -137,212 +129,35 @@ class User
     /**
      * User Class Constructor
      *
-     * @param \g7mzr\db\interfaces\InterfaceDatabaseDriver $db Database Connection Object.
-     *
      * @access public
      */
-    public function __construct(\g7mzr\db\interfaces\InterfaceDatabaseDriver $db)
+    public function __construct()
     {
-        $this->db       = $db ;
-        $this->userName = "None";
+    }
+
+
+    /**
+     * User Class Load User Data
+     *
+     * @param array $userData The information on the current user.
+     *
+     * @return void
+     * @access public
+     */
+    public function loadUserData(array $userData)
+    {
+        $this->userName = $userData['userName'];
+        $this->userId = $userData['userId'];
+        $this->realName = $userData['realName'];
+        $this->userEmail = $userData['userEmail'];
+        $this->displayrows = $userData['displayRows'];
+        $this->last_seen_date = $userData['last_seen_date'];
+        $this->passwdAgeMsg = $userData['passwdAgeMsg'];
+        $this->theme = $userData['theme'];
+        $this->zoomText = $userData['zoomText'];
+        $this->encryptedPasswd = $userData['encryptedPasswd'];
+        $this->permissions = $userData['permissions'];
         $this->styleDir = __DIR__ . "/style/";
-    }
-
-    /**
-     * Login and register the current user.  Once a user is logged in only the
-     * register function needs to be called.
-     *
-     * @param string  $userName  Name of user to be logged in.
-     * @param string  $passwd    Password for user.
-     * @param integer $passwdage Control Parameter for password aging.
-     * @param array   $config    Array containing the site preferences.
-     *
-     * @return mixed True if user logged in.  Returns WEBTEMPLATE::Error if unable
-     *               to complete login
-     *               error code 1 = system error, 2 = account locked,
-     *               3 = invalid username/password, 4 = unable to register user
-     *
-     * @access public
-     */
-    final public function login(
-        string $userName,
-        string $passwd,
-        int $passwdage,
-        array $config
-    ) {
-        //date_default_timezone_set('Europe/London');
-        // Set Passwd Age Msg to default value
-        $this->passwdagemsg = '';
-
-        // Set up the default variables
-        // User not logged in
-        $loggedin = false;
-
-        // Transfer user name to local variable
-        $tempuser = $userName;
-
-        // Default WEBTEMPLATE Error Code
-        $errorCode = 1;
-
-        $fields = array(
-            'user_name',
-            'last_seen_date',
-            'user_enabled',
-            'user_passwd',
-            'date(passwd_changed) as passwd_changed',
-            'last_failed_login'
-        );
-        $searchData = array (
-            'user_name' => $tempuser,
-        );
-
-        $userArray = $this->db->dbselectsingle('users', $fields, $searchData);
-        // Test if data has been returned
-        if (\g7mzr\db\common\Common::isError($userArray)) {
-            if ($userArray->getCode() == DB_ERROR_NOT_FOUND) {
-                // Invalid Username and password
-                $errorMsg = gettext("Invalid Username and password");
-                $errorCode = 3;
-            } else {
-                // Error Running SQL Query
-                $errorMsg = gettext("Error Running SQL Query");
-                $errorCode = 1;
-            }
-            return\g7mzr\webtemplate\general\General::raiseError($errorMsg, $errorCode);
-        }
-
-        // Test if the correct password has been entered
-        $userpwd = chop($userArray['user_passwd']);
-        if (\g7mzr\webtemplate\general\General::verifyPasswd($passwd, $userpwd) == false) {
-            // Set the Last invalid Password field
-            $insertData = array('last_failed_login' => 'now()');
-            $searchData = array('user_name' => $userArray['user_name']);
-            $result = $this->db->dbupdate('users', $insertData, $searchData);
-
-            // Return Error Message
-            $errorMsg = gettext("Invalid Username and password");
-            $errorCode = 1;
-            return\g7mzr\webtemplate\general\General::raiseError($errorMsg, $errorCode);
-        }
-
-        // Check is is more than x seconds since the last failed login
-        $last_failed_login = strtotime(chop($userArray['last_failed_login']));
-
-        if (($last_failed_login + $this->failedLoginTimeout) > time()) {
-            // Set the Last invalid Password field
-            $insertData = array('last_failed_login' => 'now()');
-            $searchData = array('user_name' => $userArray['user_name']);
-            $result = $this->db->dbupdate('users', $insertData, $searchData);
-
-            // Return Error Message
-            $errorMsg = gettext("Invalid Username and password");
-            $errorCode = 1;
-            return\g7mzr\webtemplate\general\General::raiseError($errorMsg, $errorCode);
-        }
-
-        // Check if the account has been locked.
-        $userEnabled = chop($userArray['user_enabled']);
-        if ($userEnabled == 'N') {
-            // Account Locked.
-            $errorMsg = gettext("Invalid Username and password");
-            $errorCode = 2;
-            return\g7mzr\webtemplate\general\General::raiseError($errorMsg, $errorCode);
-        }
-
-        // User is logged in
-        $loggedin = true;
-        $this->last_seen_date = chop($userArray['last_seen_date']);
-        $changedate = strtotime(chop($userArray['passwd_changed']));
-
-        $result = $this->passwordAge($passwdage, $changedate);
-
-        if (\g7mzr\webtemplate\general\General::isError($result)) {
-            return $result;
-        }
-        // Logged in. Register the user
-        $registeredOk = $this->register($userName, $config);
-        if (!\g7mzr\webtemplate\general\General::isError($registeredOk)) {
-            return true;
-        } else {
-            // Failed to register the user okay
-            $errorMsg = gettext("Unable to Complete Login Process");
-            return\g7mzr\webtemplate\general\General::raiseError($errorMsg, 4);
-        }
-    }
-
-
-
-    /**
-     * Register a currently logged in user.  Store their realname, userid,
-     * encrypted password, permissions and preferences in the USER Class object.
-     *
-     * @param string $userName    Name of user logged in.
-     * @param array  $preferences Array containing the site preferences.
-     *
-     * @return mixed True if user registered.  Returns WEBTEMPLATE::Error on error.
-     *
-     * @access public
-     */
-    final public function register(string $userName, array $preferences)
-    {
-
-        // Flush all data
-        $this->userName = "";
-        $this->realName = "None";
-        $this->userEmail = "";
-        $this->userId = 0;
-        $this->encryptedPasswd = '';
-        $this->permissions = array();
-        $this->theme = '';
-        $this->zoomText = false;
-        $this->displayrows = '1';
-
-        // Set local variables
-        $registeredOk = true;
-
-        // Save usernaem in the Class vairable
-        $this->userName = $userName;
-
-        // Get the users details from the database
-        $fieldNames = array(
-            'user_id',
-            'user_name',
-            'user_passwd',
-            'user_realname',
-            'user_email',
-            'user_enabled',
-            'user_disable_mail',
-            'date(last_seen_date)'
-        );
-        $searchData = array('user_name' => $this->userName);
-        $userArray = $this->db->dbselectsingle('users', $fieldNames, $searchData);
-        if (\g7mzr\db\common\Common::isError($userArray)) {
-            // Error encountered when selecting user from database
-            // Registration process has failed
-            $msg = gettext("Unable to Register User");
-            return\g7mzr\webtemplate\general\General::raiseError($msg, 1);
-        }
-
-        // Populate the class variables with the collected data
-        $this->realName = $userArray['user_realname'];
-        $this->userEmail = $userArray['user_email'];
-        $this->userId = $userArray['user_id'];
-        $this->encryptedPasswd = $userArray['user_passwd'];
-
-        // Get the Users Preferences if they have any.
-        $this->getPreferences($preferences);
-
-        // Failure to set last seen date fails registration
-        $insertData = array('last_seen_date' => 'now()');
-        $searchData = array('user_id' => $this->userId);
-        $result = $this->db->dbupdate('users', $insertData, $searchData);
-        if (\g7mzr\db\common\Common::isError($result)) {
-            // Error encountered when selecting user from database
-            // Registration process has failed
-            $msg = gettext("Unable to Register User");
-            return\g7mzr\webtemplate\general\General::raiseError($msg, 1);
-        }
-        return true;
     }
 
 
@@ -503,184 +318,7 @@ class User
      */
     final public function getPasswdAgeMsg()
     {
-        return $this->passwdagemsg;
-    }
-
-    /**
-     * Function to get user's preferences from the database
-     *
-     * This function gets the users preferences from the database.  If no preferences
-     * are found or there is a database error the default preferences are used.
-     *
-     * @param array $preferences Array containing the site preferences.
-     *
-     * @return boolen True
-     *
-     * @access private
-     */
-    private function getPreferences(array $preferences)
-    {
-
-        // Set the users preferences
-        // Set the default Site preferences for user.
-        $this->theme = $preferences['theme']['value'];
-        $this->zoomText = $preferences['zoomtext']['value'];
-        $this->displayrows = $preferences['displayrows']['value'];
-
-        // Search database for User Preferences
-        $fieldNames = array('settingname', 'settingvalue');
-        $searchData = array('user_id' => $this->userId);
-        $userArray = $this->db->dbselectmultiple(
-            'userprefs',
-            $fieldNames,
-            $searchData
-        );
-
-        // Missing Preferences do not stop registration
-        if (!\g7mzr\webtemplate\general\General::isError($userArray)) {
-            foreach ($userArray as $value) {
-                // Walk through the results
-                // Get theme
-                $this->getTheme($value, $preferences['theme']['enabled']);
-
-                // Check zoom text
-                $this->getZoomArea($value, $preferences['zoomtext']['enabled']);
-
-                // Check Display rows
-                $this->getRows(
-                    $value,
-                    $preferences['displayrows']['enabled']
-                );
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Function to get the user's preferred theme.
-     * It is called by getPreferences
-     *
-     * @param array   $value   The array holding the users preferences.
-     * @param boolean $enabled True if user selectable theme is enabled.
-     *
-     * @return boolean True in all cases
-     */
-    private function getTheme(array $value, bool $enabled)
-    {
-        // Check the theme preference
-        if ((chop($value['settingname']) == 'theme') and ($enabled == true)) {
-            // The user has set their own theme
-            // Get the user chosen theme
-            $localTheme = chop($value['settingvalue']);
-
-            // The file name for the themes main file
-            $testfile = $this->styleDir . $localTheme . '/main.css';
-
-            if (file_exists($testfile)) {
-                // The theme exists.
-                // Set it as the users chosen theme
-                $this->theme = $localTheme;
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Function to get the user's text area zoom flag
-     * It is called by getPreferences
-     *
-     * @param array   $value   The array holding the users preferences.
-     * @param boolean $enabled True if user selectable theme is enabled.
-     *
-     * @return boolean True in all cases
-     */
-    private function getZoomArea(array $value, bool $enabled)
-    {
-        if ((chop($value['settingname']) == 'zoomtext') and ($enabled == true)) {
-            // The user has set their own zoon text.
-            if (chop($value['settingvalue']) == 'true') {
-                $this->zoomText = true;
-            } else {
-                $this->zoomText = false;
-            }
-        }
-
-        return true;
-    }
-    /**
-     * Function to get the user's preferred theme.
-     * It is called by getPreferences
-     *
-     * @param array   $value   The array holding the users preferences.
-     * @param boolean $enabled True if user selectable theme is enabled.
-     *
-     * @return boolean True in all cases
-     */
-    private function getRows(array $value, bool $enabled)
-    {
-        if ((chop($value['settingname']) == 'displayrows') and ($enabled == true)) {
-            // The user has set the number of sata rows to display
-            $this->displayrows = chop($value['settingvalue']);
-        }
-
-        return true;
-    }
-
-
-    /**
-     * Function to test if a users password has expired
-     *
-     * This function checks if a users password has expired
-     *
-     * @param integer $passwdage  Control Parameter for password aging.
-     * @param integer $changedate The date the users password was last changed.
-     *
-     * @return mixed true if password does not need immediate changing
-     *                    or WEBTEMPLATE Error object if the password must be changed
-     *
-     * @access private
-     */
-    private function passwordAge(int $passwdage, int $changedate)
-    {
-        switch ($passwdage) {
-            case 2:
-                $passwddays = 60;
-                break;
-            case 3:
-                $passwddays = 90;
-                break;
-            case 4:
-                $passwddays = 180;
-                break;
-            default:
-                $passwddays = 0;
-                break;
-        }
-
-        if ($passwddays == 0) {
-            return true;
-        }
-
-        $todaysdate = time();
-        $secs = $todaysdate - $changedate;
-        $days = (int) ($secs / 86400);
-        $daysleft = $passwddays - $days;
-        if (($daysleft > 0) and ($daysleft < 15)) {
-            $str1 = gettext("Your password is due to expire in ");
-            $str2 = gettext(" days. To change your password click ");
-            $str3 = '<a href="/userprefs.php?tab=account">here.</a>';
-            $this->passwdagemsg = $str1 . $daysleft . $str2 . $str3;
-        }
-        if (($daysleft < 1) or ($changedate == null)) {
-            $errorMsg = gettext("Your password has expired and must be changed");
-            $errorCode = 5;
-            return\g7mzr\webtemplate\general\General::raiseError(
-                $errorMsg,
-                $errorCode
-            );
-        }
-        return true;
+        return $this->passwdAgeMsg;
     }
 
     /**
