+++ tests/rest/RestGroupsTest.php
@@ -129,9 +129,11 @@ class RestGroupsTest extends TestCase
     public function testGroupPermissionsLoggedinNoAccess()
     {
         // Register the user to look as if they are logged in
-        $registered = $this->webtemplate->user()->register(
+        $userData = array();
+        $registered = $this->webtemplate->register()->register(
             'secnone',
-            $this->webtemplate->config()->read('pref')
+            $this->webtemplate->config()->read('pref'),
+            $userData
         );
         if (\g7mzr\webtemplate\general\General::isError($registered)) {
             $this->fail('Failed to register user for ' . __METHOD__);
@@ -139,12 +141,12 @@ class RestGroupsTest extends TestCase
 
         // Get the users groups
         $this->webtemplate->usergroups()->loadusersgroups(
-            $this->webtemplate->user()->getUserId()
+            $userData['userId']
         );
 
         $this->webtemplate->session()->createSession(
             'secnone',
-            $this->webtemplate->user()->getUserId(),
+            $userData['userId'],
             false
         );
 
@@ -175,9 +177,11 @@ class RestGroupsTest extends TestCase
     {
 
         // Register the user to look as if they are logged in
-        $registered = $this->webtemplate->user()->register(
+        $userData = array();
+        $registered = $this->webtemplate->register()->register(
             'secgroups',
-            $this->webtemplate->config()->read('pref')
+            $this->webtemplate->config()->read('pref'),
+            $userData
         );
         if (\g7mzr\webtemplate\general\General::isError($registered)) {
             $this->fail('Failed to register user for ' . __METHOD__);
@@ -185,12 +189,12 @@ class RestGroupsTest extends TestCase
 
         // Get the users groups
         $this->webtemplate->usergroups()->loadusersgroups(
-            $this->webtemplate->user()->getUserId()
+            $userData['userId']
         );
 
         $this->webtemplate->session()->createSession(
             'secgroups',
-            $this->webtemplate->user()->getUserId(),
+            $userData['userId'],
             false
         );
