+++ tests/rest/RestVersionTest.php
@@ -122,9 +122,11 @@ class RestVersionTest extends TestCase
     public function testVersionGetLoggedIn()
     {
         // Register the user to look as if they are logged in
-        $registered = $this->webtemplate->user()->register(
+        $userData = array();
+        $registered = $this->webtemplate->register()->register(
             'phpunit',
-            $this->webtemplate->config()->read('pref')
+            $this->webtemplate->config()->read('pref'),
+            $userData
         );
         if (\g7mzr\webtemplate\general\General::isError($registered)) {
             $this->fail('Failed to register user for ' . __METHOD__);
@@ -132,7 +134,7 @@ class RestVersionTest extends TestCase
 
         // Get the users groups
         $this->webtemplate->usergroups()->loadusersgroups(
-            $this->webtemplate->user()->getUserId()
+            $userData['userId']
         );
 
         $method = "GET";
